events-api/requirements.txt
alembic==1.4.2
astroid==2.4.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
PyJWT==1.7.1
pylint==2.5.0
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.0
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
